backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
python-dotenv>=1.2.0
